Update Portrait decleration + Update demo

+++ src/demo/platformer/box.hpp
@@ -5,7 +5,7 @@
 class Box : public mia::WorldObject
 {
 public:
-    Box(int x, int y, int sx = 1, int sy = 1);
+    Box(float x, float y, float sx = 1, float sy = 1);
 
     void Update(uint8_t message) override;
 };
